[#4549] Replace view with router to handle details overview
- Change subdomain wcaro to rsr2.
- Create separate components for indicators.
