+++ akvo/rsr/dir/app/modules/wcaro-index/components/Summarize.jsx
@@ -1,6 +1,7 @@
 /* eslint-disable react/no-danger */
 import React from 'react'
 import { Row, Col, Icon, Typography } from 'antd'
+import { Link } from 'react-router-dom'
 
 const { Title } = Typography
 
@@ -14,6 +15,9 @@ export const Summarize = ({ items }) => {
             {item.amount ? item.amount.toLocaleString('en-US') : ''}
             <div dangerouslySetInnerHTML={{ __html: item.label }} />
           </Title>
+          <Link to={`/insight/${index}`}>
+            See More
+          </Link>
         </Col>
       ))}
     </Row>
