+++ akvo/rsr/dir/app/modules/wcaro-index/views/Insight.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+const Insight = () => {
+  return (
+    <>
+      <h1>Insight</h1>
+      <Link to="/">Go back</Link>
+    </>
+  )
+}
+
+export default Insight
